added DashboardController.js and putted index there
