+++ src/routes.js
@@ -2,9 +2,10 @@ const express = require("express");
 const routes = express.Router();
 const ProfileController = require("./controllers/ProfileController");
 const JobController = require("./controllers/JobController");
+const DashboardController = require("./controllers/DashboardController");
 
 // request e response
-routes.get('/', JobController.index);
+routes.get('/', DashboardController.index);
 routes.get('/job', JobController.create);
 routes.post('/job', JobController.save);
 routes.get('/job/:id', JobController.show);
